pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-chess-coach"
version = "0.2.0"
description = "Agentic AI chess coach with OpenAI Agents SDK tools, game simulation, and PGN export"
requires-python = ">=3.10"
dependencies = [
  "python-chess>=1.999",
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "openai-agents>=0.0.12",
  "pytest"
]

[tool.pytest.ini_options]
pythonpath = ["."]
